Revert "refactor: Merged EnhancedMarkdown.vue & Markdown.vue (#12214)"
This reverts commit 8b1ceb836ba643553fe0da379285ba92c88891fe.
